refactor(halo/portal): support explicit destinations and shards (#1194)
Extends `protal` module in halo to support explicit `destChainID` and
`ShardIDs`. This is required to make the portal module "future-proof" in
order to support non-broadcast type xmsgs.

Non-broadcast xmsgs from halo cannot share shards with broadcast xmsgs
since this would result in gaps in xmsg offsets of broadcast messages
when submitting to some portals.

task: none

+++ relayer/app/types.go
@@ -43,7 +43,7 @@ func submissionToBinding(sub xchain.Submission) bindings.XSubmission {
 		msgs = append(msgs, bindings.XMsg{
 			SourceChainId: msg.SourceChainID,
 			DestChainId:   msg.DestChainID,
-			ShardId:       msg.ShardID,
+			ShardId:       uint64(msg.ShardID),
 			Offset:        msg.StreamOffset,
 			Sender:        msg.SourceMsgSender,
 			To:            msg.DestAddress,
